[CHORE] : moved all type definition in Layout.tsx, and jeekim written files to type.tsx
